backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.95.0",
    "fastapi>=0.135.3",
    "google-cloud-bigquery>=3.41.0",
    "uvicorn>=0.44.0",
]
